requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
plotly>=5.17.0
matplotlib>=3.7.0

# Dependencias del proyecto original
pytesseract>=0.3.10
PyMuPDF>=1.23.0
transformers>=4.35.0
torch>=2.0.0
openai>=1.3.0
pyodbc>=5.0.0